tests: remove symlinking shenanigans in favour of two copies
